scanner/requirements.txt
azure-identity==1.15.0
azure-mgmt-storage==21.0.0
azure-mgmt-network==25.0.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-compute==30.0.0
azure-mgmt-resource==23.0.0
anthropic==0.25.0
prometheus-client==0.19.0
python-dotenv==1.0.0

